requirements.txt
envyaml==1.10.211231
Jinja2==3.0.1
paramiko==2.11.0
python-dotenv==0.19.0
PyYAML==5.4.1
requests==2.26.0
